+++ packages/ApiTests/src/package-test.ts
@@ -26,6 +26,7 @@ import './dapi/fetch';
 import './dapi/groups';
 import './dapi/dapi';
 import './dapi/domains';
+import './dapi/domains-batch';
 import './dapi/connection';
 import './dapi/entities';
 import './dapi/entities-save';